+++ src/main/java/com/lostjs/wx4j/data/response/GroupMember.java
@@ -14,7 +14,7 @@ public class GroupMember {
 
     private String nickName;
 
-    private int attrStatus;
+    private long attrStatus;
 
     private String pyInitial;
 
@@ -57,12 +57,12 @@ public class GroupMember {
         this.nickName = WxNickNameConverter.convertFromHtml(nickName);
     }
 
-    public int getAttrStatus() {
+    public long getAttrStatus() {
         return attrStatus;
     }
 
     @JsonSetter("AttrStatus")
-    public void setAttrStatus(int attrStatus) {
+    public void setAttrStatus(long attrStatus) {
         this.attrStatus = attrStatus;
     }
